requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
python-dotenv